Fixed missing character
Problem was that pre-3.0 Python reads a line, and the line ends with
“\r\n”. post-3.0 Python ends a readline with “\n”. This meant that all
of my references of [:-2] were 1 character too much depending on the
version of python.
utils.py now checks for version once a line containing the delimiter
has been found. If the version is before 3.0, is removes two
characters, appends a “\n”, and saves that string back into itself.
final string is then returned.