Removed files. Updates. Problem...
- removed old testing files (no longer needed)

utils.py
- renamed checkVersion to isLessThanVersion (which makes a lot more
sense)

SpeedTestFile
- changed f to fs (for file stream. sorry. OCD)

STDs
- Removed adding to structure parts, put into function (addToStructure)
- added __str__ methods to STDs
- “-cs” option now works

IndividualSpeedTest
- reworked to use data stream

!!! IMPORTANT PROBLEM !!!!
Run main.py with the option -cs (after renaming the variable DataRoot
in STDs. DataRoot is near the beginning of the class)
There seems to be an issue where Individual tests are added to ALL
objects. After running main.py with the “-cs” option, you should see
that each test has 36 tests. This is wrong!!! Maybe we do need to go
back to a string and not a data stream.