Both of our output functions should be working
At this point, both functions (the one that produces the distribution,
and the one that produces the updated master CSV file) should be able
to run with no errors in Python3.

Next to add is the check for tests that had errors, but this may
already be accounted for in our program.