Converting Indiv to str parsing. created TCP/UDP
This is a step in the conversion of IndivSpeedTest into two separate
classes (TCP and UDP) and having them parse the data from a string
rather than a file stream

IndivSpeedTest
- simplified class. Only has base variables, a simple init, and string
just returns the string that was passed in. Does NO PARSING

TCP
- Currently looks exactly like old IndivSpeedTest
- Currently has 1 new class variable, WindowSize

UDP
- Currently looks exactly like old IndivSpeedTest
- Currently has 2 new class variables, DatagramSize and TargetBandwidth

Pings, PingThread, STDs, STFile
- made imports of global_str_padding to be one line (use of ‘;’ )
- all class variables of “this_” variety are renamed to “my”. Not as
hard to type

STFile
- End of loadHeaderInfo() now calls parseIndivTests(), and passes in
the remaining file contents (fs.read())
- parseIndivTests takes the string, and splits it based on ‘\n\n’. It
then iterates through the “chunks”, and appends chunks that are a part
of a single test. These new “chunks” are put into sortedFileContents.
The elements in this array are then analyzed (to determine TCP or
UDP-ness), and the appropriate constructor is called.