fixing of appending to results CSV
The program flow is such that, in STFile, we first see if the file is
recorded in the master CSV.

If what was returned was an index, it creates an empty array. It then
gets all of the TCP threads in the correct order (which is test # 1, 2,
4, and then 5). If the test exists, it runs the test’s function. If the
test does not exist, the mini-array appended for that test is 4 values
of error.

In the TCP_Test, the function to get the stdev and median values checks
if there are errors in the test. If there are, a 4 element array of
“None” is returned. Otherwise, a 4 element array of the stdev and
median of the up and down streams are returned.

After the 4 element array is returned, we extend the end array with the
values, and move to the next test. The 16 value array at the end of
this is appended to the correct CSV array for that file.