Most of the changes were to improve how the tests were parsed, stored in the STDS, and how values were appended to the CPUC_Results CSV.
IndivSpeedTest
- Checking for Errors now correctly accounts for possible error messages
- Sometimes no iPerfCommand was issued before the tests were quitted.
This is now accounted for

TCP_Test
- put thread splitting part of createPingThreads into try/except block.
If something in it broke, there must have been an error in what was
printed out by iPerf. The test is then ignored.
- array of 4 values returned by create_Array_For_Results now contains
values that are a little more descriptive

STDS
- created array of bad_info_Files (to hold file objects whose data is
not very correct
- appending StDev/Median to CSV now appends the objects bad_info_Files.
- At the end of the function, if there are still empty values in the
last 16 cells (where StDev and Median go), then it is filled with
“FileMissingError”.
- __str__ prints out the info of the objects in bad_info_Files is
recursivelyPrint is True

SpeedTestFile
- the function that search for the row index of the file in the
CPUC_Results CSV now tries with locationID, Date, and Time if index was
still None at the end of the first loop