Many much changes. Am confused?
main
- Added 5th option (tcp throughput)
- check reading choice, choices 3+ go into one loop. This loop asks for
the original csv file to be chosen, import it, and then determines
which function to do (3, 4, or 5)
- STDs now have a single function for all 3 analysis, and the function
takes a reference to the CSV, the csv file name, an array of the
headers that will be added to the top of the csv, and a string (the
function that will be run).

STDS
- now 1 function for (currently) all 3 analyses. Function appends
headers, determines which function will be run (in SpeedTestFile), and
then loops through all of the files parsed, running the chosen
function.

SpeedTestFile
- changed how Ping tests are determined to be Ping tests from text
(there was a case of one not being categorized as a Ping test, as it
was missing the string we were looking for)
- - - also, Ping test now correctly casts it’s parsed values (there was
a small section where the parsed string was not cast as a float)
- creation of TCP_throughput calculations and appending
- - first get east/west ping, get RTT average and use new utils
function to get theoretical TCP_TP. Then, for each TCP test, compare
theoretical TP to measured TP for UP and DOWN direction separately.

utils
- function to calculate TCP throughput. inputs must be in bytes,
milliseconds, and percent. The values are then converted to Kbits,
seconds, and decimal (respectively). We then plug the numbers into the
function.